voice_agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
numpy==2.2.6
openai==2.16.0
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
